Proof-of-concept: Animated scrolling
This is a proof-of-concept patch to make scrolled windows animate
motion. The implementation of the tick callback is in GtkAdjustment,
so we can animate both scrollbar clicks and key bindings.

Still to do: figure out what to do about duration - does it need to
depend on the distance of the jump ? Is the easing function the right
choice ? Do we want to animate all range jumps, or just scrollbars ?

https://bugzilla.gnome.org/show_bug.cgi?id=732376