cssshadowvalue: Convert the current point into a translate
Pango uses the current point to determine where to paint the text, which
is considerably inconvenient since it doesn't work well with
translations.

To make this simpler for us, turn the current point into a translation,
which lets us merge the three different pieces of context setup into
just one.